1.调整代码生成器和模板 2.利用代码生成器生成组织机构模块

+++ generate-code/src/main/java/mofa/wangzhe/code/StaticCode.java
@@ -92,21 +92,26 @@ public class StaticCode {
         path = path + fileName;
         File file2 = new File(path);
         if (file2.exists()) {
-            boolean delete = file2.delete();
-            if (delete) {
-                boolean newFile = false;
-                try {
-                    newFile = file2.createNewFile();
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-                if (!newFile) {
-                    log.info("创建文件失败");
+            if (staticCodeModel.isReplace()) {
+                boolean delete = file2.delete();
+                if (delete) {
+                    boolean newFile = false;
+                    try {
+                        newFile = file2.createNewFile();
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                    }
+                    if (!newFile) {
+                        log.info("创建文件失败");
+                        System.exit(-1);
+                    }
+                } else {
+                    log.info("删除文件失败");
                     System.exit(-1);
                 }
             } else {
-                log.info("删除文件失败");
-                System.exit(-1);
+                log.info("文件已存在，不在生成");
+                return;
             }
         }
